renaming components file name + uniting stores
