+++ src/components/Pixel/Pixel.jsx
