+++ src/components/ResolutionSlider/ResolutionSlider.jsx
