+++ src/redux/reducers/boardSlice.ts
@@ -1,21 +1,18 @@
 import { createSlice } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
-
-export interface Pixel {
-  color: string;
-}
+import BoardOptions from "../../components/BoardOptions/BoardOptions";
 
 export interface BoardState {
-  pixels: Pixel[];
-}
-
-export interface ISetPixelColorByIndexPayload {
-  index: number;
-  color: string;
+  pixels: { color: string }[];
+  boardOptions: {
+    gridResolution: number;
+    isShowingGrid: boolean;
+  };
 }
 
 const initialState: BoardState = {
   pixels: Array(16 ** 2).fill({ color: "#FFF" }),
+  boardOptions: { gridResolution: 16, isShowingGrid: false },
 };
 
 export const boardSlice = createSlice({
@@ -24,18 +21,29 @@ export const boardSlice = createSlice({
   reducers: {
     initializePixels: (state, action: PayloadAction<number>) => {
       state.pixels = Array(action.payload ** 2).fill({ color: "#FFF" });
-      console.log(state.pixels);
     },
     setPixelColorByIndex: (
       state,
-      action: PayloadAction<ISetPixelColorByIndexPayload>,
+      action: PayloadAction<{ index: number; color: string }>,
     ) => {
       const { color, index } = action.payload;
       if (state.pixels[index]) state.pixels[index]!.color = color;
     },
+    setGridResolution: (state, action: PayloadAction<number>) => {
+      state.boardOptions.gridResolution = action.payload;
+    },
+    toggleShowingGrid: (state) => {
+      const { boardOptions } = state;
+      boardOptions.isShowingGrid = !boardOptions.isShowingGrid;
+    },
   },
 });
 
-export const { initializePixels, setPixelColorByIndex } = boardSlice.actions;
+export const {
+  initializePixels,
+  setPixelColorByIndex,
+  setGridResolution,
+  toggleShowingGrid,
+} = boardSlice.actions;
 
 export default boardSlice.reducer;
