+++ src/redux/store.ts
@@ -1,12 +1,10 @@
 import { configureStore } from "@reduxjs/toolkit";
-import boardOptionsSlice from "./reducers/boardOptionsSlice";
 import boardSlice from "./reducers/boardSlice";
 import drawingOptionsSlice from "./reducers/drawingOptionsSlice";
 
 export const store = configureStore({
   reducer: {
     drawingOptions: drawingOptionsSlice,
-    boardOptions: boardOptionsSlice,
     board: boardSlice,
   },
 });
